context.js becomes typescript
