--- app/context.js
@@ -1,46 +1,0 @@
-import {createContext, useEffect, useState} from "react";
-import {auth} from '@/lib/firebase';
-import {onAuthStateChanged} from 'firebase/auth';
-
-export const AuthContext = createContext();
-
-//provider
-export function AuthProvider({children}) {
-    const [user, setUser] = useState(null);
-
-    useEffect(() => {
-        const unsubscribe = onAuthStateChanged(auth, (user) => {
-            if (user) {
-                setUser(user);
-            } else {
-                setUser(null);
-            }
-        });
-        return () => unsubscribe();
-    }, [auth]);
-
-    return (
-        <AuthContext.Provider value={user}>
-            {children}
-        </AuthContext.Provider>
-    )
-}
-
-export function useAuth() {
-    const [user, setUser] = useState(null);
-    const [loading, setLoading] = useState(true);
-
-    useEffect(() => {
-        const unsubscribe = onAuthStateChanged(auth, (user) => {
-            if (user) {
-                setUser(user);
-            } else {
-                setUser(null);
-            }
-            setLoading(false);
-        });
-        return () => unsubscribe();
-    }, [auth]);
-
-    return {user, loading};
-}
